// src/app/page/master/ruangan/ruangan.interface.ts
import { EventEmitter } from '@angular/core';

export interface Ruangan {

	alamatEmail?: string;
	faksimile?: string;
	fixedPhone?: string;
	jamBuka?: string;
	jamTutup?: string;
	kdDepartemen?: string;
	kdKelasHead?: string;
	kdModulAplikasi?: string;
	kdPegawaiKepala?: string;
	kode?: string;
	kodeExternal?: string;
	lokasiRuangan?: string;
	mobilePhone?: string;
	namaExternal?: string;
	namaRuangan?: string;
	noCounter?: any;
	noRuangan?: string;
	prefixNoAntrian?: string;
	reportDisplay?: string;
	statusEnabled?: true;
	statusViewData?: true;
	version?: any;
	website?: string;
	noRec?: string;
	label?: any;
	value?: any;
	kdRuanganHead?:any;

}